requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.8.3
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
